+++ ricaun.Revit.UI.Tests/Panels/RevitPanelMoveTests.cs
@@ -1,10 +1,12 @@
-﻿using NUnit.Framework;
+﻿using Autodesk.Revit.UI;
+using NUnit.Framework;
 using System.Linq;
 
 namespace ricaun.Revit.UI.Tests.Panels
 {
     public class RevitPanelMoveTests : BaseCreatePanelTests
     {
+        protected override string PanelName => "Move" + base.PanelName;
         private const string TabIdAddIns = "Add-Ins";
         private const string TabIdModify = "Modify";
 
@@ -19,6 +21,15 @@ namespace ricaun.Revit.UI.Tests.Panels
         {
             ribbonPanel.MoveToRibbonTab();
             Assert.AreEqual(TabIdModify, ribbonPanel.GetRibbonTab().Id);
+        }
+
+        [Test]
+        public void MovePanelTo_Modify_CreatePanel_SameName(UIControlledApplication application)
+        {
+            ribbonPanel.MoveToRibbonTab();
+            var panel = application.CreatePanel(PanelName);
+            panel.Remove();
+            Assert.AreEqual(panel.Name, ribbonPanel.Name);
         }
 
         [Test]